// src/app/interfaces/sysrecord.ts
export interface AdditionalPrivilege {
    id: number;
    uploadDate: Date;
    hospitalCode: string;
    privilegeGroupCode: string;
    pdfLocation: string;
    pdfUri: string;
    remark: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    updateLocation: string;
}
// ...
export interface Hospital {
    hospitalCode: string;
    hospitalName: string;
    hospitalShortName: string;
}
// ...
export interface Privilegeroup {

    privilegeCode: string;
    privilegeDesc: string;
    privilegeType: string;
}